fix: Update README and env var handling
Due to a change in Omnivore's web page, the only way to locate the
userid is now in the body of the webhook. This commit updates the readme
to reflect the new method of finding the userid and changes the check
for the userid variable from a panic to a log message so the server can
be used for this purpose.